Rename filter_traits to type_traits
