+++ crates/minikalman/src/buffer_types/innovation_vector_buffer.rs
@@ -1,8 +1,8 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
 
-use crate::filter_traits::InnovationVector;
 use crate::matrix_traits::{Matrix, MatrixMut};
+use crate::type_traits::InnovationVector;
 use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 
 pub struct InnovationVectorBuffer<const MEASUREMENTS: usize, T, M>(M, PhantomData<T>)
