+++ crates/minikalman/src/buffer_types/kalman_gain_matrix_buffer.rs
@@ -1,8 +1,8 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
 
-use crate::filter_traits::KalmanGainMatrix;
 use crate::matrix_traits::{Matrix, MatrixMut};
+use crate::type_traits::KalmanGainMatrix;
 use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 
 pub struct KalmanGainMatrixBuffer<const STATES: usize, const MEASUREMENTS: usize, T, M>(
