+++ crates/minikalman/src/buffer_types/measurement_process_noise_covariance_matrix_buffer.rs
@@ -1,5 +1,5 @@
-use crate::filter_traits::MeasurementProcessNoiseCovarianceMatrix;
 use crate::matrix_traits::{Matrix, MatrixMut};
+use crate::type_traits::MeasurementProcessNoiseCovarianceMatrix;
 use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
