+++ crates/minikalman/src/buffer_types/state_prediction_vector_buffer.rs
@@ -1,8 +1,8 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
 
-use crate::filter_traits::StatePredictionVector;
 use crate::matrix_traits::{Matrix, MatrixMut};
+use crate::type_traits::StatePredictionVector;
 use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 
 pub struct StatePredictionVectorBuffer<const STATES: usize, T, M>(M, PhantomData<T>)
