+++ crates/minikalman/src/buffer_types/system_covariance_matrix_buffer.rs
@@ -1,8 +1,8 @@
 use core::marker::PhantomData;
 use core::ops::{Index, IndexMut};
 
-use crate::filter_traits::SystemCovarianceMatrix;
 use crate::matrix_traits::{Matrix, MatrixMut};
+use crate::type_traits::SystemCovarianceMatrix;
 use crate::{IntoInnerData, MatrixData, MatrixDataMut, MatrixDataOwned};
 
 pub struct SystemCovarianceMatrixBuffer<const STATES: usize, T, M>(M, PhantomData<T>)
